use GRPC http and remove server stuff
