+++ src/Speckle.Logging/TraceBuilder.cs
@@ -1,4 +1,6 @@
 using OpenTelemetry;
+using OpenTelemetry.Exporter;
+using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
 
 namespace Speckle.Logging;
@@ -7,13 +9,27 @@ public class TraceBuilder(IDisposable? traceProvider) : IDisposable
 {
   public static IDisposable Initialize(string application, SpeckleLogConfiguration logConfiguration)
   {
-    var tracerProviderBuilder = Sdk.CreateTracerProviderBuilder().AddSource(application).AddZipkinExporter();
-
+    var tracerProviderBuilder = Sdk.CreateTracerProviderBuilder()
+      .AddSource(application)
+      .ConfigureResource(r =>
+      {
+        r.AddAttributes(new List<KeyValuePair<string, object>>
+        {
+          new("service.name", application)
+        });
+      })
+      .AddHttpClientInstrumentation();
+    if (logConfiguration.LogToOtel)
+    {
+      tracerProviderBuilder = tracerProviderBuilder.AddOtlpExporter(x =>
+      {
+        x.Protocol = OtlpExportProtocol.HttpProtobuf;
+      });
+    }
     if (logConfiguration.LogToConsole)
     {
       tracerProviderBuilder = tracerProviderBuilder.AddConsoleExporter();
     }
-
     return new TraceBuilder(tracerProviderBuilder.Build());
   }
